+++ src/septentrio_gnss_driver/node/rosaic_node.cpp
@@ -91,8 +91,7 @@ bool rosaic_node::ROSaicNode::getROSParams()
     param("aux1_frame_id", settings_.aux1_frame_id, (std::string) "aux1");
     param("vehicle_frame_id", settings_.vehicle_frame_id, settings_.poi_frame_id);
     param("lock_utm_zone", settings_.lock_utm_zone, true);
-    getUint32Param("leap_seconds", settings_.leap_seconds,
-                           static_cast<uint32_t>(18));
+    param("leap_seconds", settings_.leap_seconds, -128);
 
     // Communication parameters
     param("device", settings_.device, std::string("/dev/ttyACM0"));